requirements.txt
selenium==4.15.2
requests==2.31.0
beautifulsoup4==4.12.2
lxml==4.9.3
fake-useragent==1.4.0
undetected-chromedriver==3.5.4
python-dotenv==1.0.0
pandas==2.1.3
schedule==1.2.0
python-dateutil==2.8.2
pytz==2023.3
webdriver-manager==4.0.1
pillow==10.1.0
opencv-python==4.8.1.78
pytesseract==0.3.10
cloudscraper==1.2.71
httpx==0.25.2
aiohttp==3.9.1
asyncio==3.4.3
setuptools==69.0.3
